[DEVOPS-694] Move redis to service folder
